requirements.txt
pandas>=1.5.0
numpy>=1.20.0
statsmodels>=0.13.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
pyyaml>=6.0
requests>=2.28.0
openpyxl>=3.0.0
pathlib2>=2.3.0
typing-extensions>=4.0.0